+++ src/main/java/step3/controller/CarRacingController.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.controller;
+
+import step3.domain.CarRacing;
 
 public class CarRacingController {
