+++ src/main/java/step3/controller/InputView.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.controller;
 
 import java.util.Scanner;
