+++ src/main/java/step3/controller/ResultView.java
@@ -1,4 +1,7 @@
-package step3;
+package step3.controller;
+
+import step3.domain.Car;
+import step3.domain.Cars;
 
 import java.util.List;
 import java.util.stream.Collectors;
