+++ src/test/java/step3/CarRacingTest.java
@@ -1,12 +1,13 @@
 package step3;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import step3.domain.Car;
+import step3.domain.CarRacing;
+import step3.domain.Cars;
 
-import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.*;
