+++ src/test/java/step3/CarUtilsTest.java
@@ -3,6 +3,7 @@ package step3;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import step3.domain.CarUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
