+++ src/test/java/step3/RacingSheetTest.java
@@ -2,6 +2,9 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.domain.Car;
+import step3.domain.Cars;
+import step3.domain.RacingSheet;
 
 import java.util.Arrays;
 import java.util.List;
